# agent-manager-service/docs/api_v1_openapi.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: Agent Manager Service API
servers:
  - url: /api/v1
paths:
...
  /orgs/{orgName}/llm-providers/{id}:
    parameters:
      - name: orgName
        in: path
        required: true
        description: Organization name/handle
        schema:
          type: string
      - name: id
        in: path
        required: true
        description: Provider UUID
        schema:
          type: string

    get:
      tags:
        - LLM Providers
      summary: Get LLM provider details
      description: Retrieve detailed information about a specific provider
      operationId: getLLMProvider
      responses:
        '200':
          description: Successfully retrieved provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LLMProviderResponse'
        '400':
          description: Invalid provider ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Provider not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - LLM Providers
      summary: Update LLM provider
      description: Update an existing LLM provider
      operationId: updateLLMProvider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLLMProviderRequest'
      responses:
        '200':
          description: Provider updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LLMProviderResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Provider not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - LLM Providers
      summary: Delete LLM provider
      description: Delete an LLM provider
      operationId: deleteLLMProvider
      responses:
        '204':
          description: Provider deleted successfully
        '400':
          description: Invalid provider ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Provider not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /orgs/{orgName}/projects/{projName}/llm-proxies/{id}:
    parameters:
      - name: orgName
        in: path
        required: true
        description: Organization name/handle
        schema:
          type: string
      - name: projName
        in: path
        required: true
        description: Project name
        schema:
          type: string
      - name: id
        in: path
        required: true
        description: Proxy UUID
        schema:
          type: string

    get:
      tags:
        - LLM Proxies
      summary: Get LLM proxy details
      description: Retrieve detailed information about a specific proxy
      operationId: getLLMProxy
      responses:
        '200':
          description: Successfully retrieved proxy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LLMProxyResponse'
        '400':
          description: Invalid proxy ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Proxy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - LLM Proxies
      summary: Update LLM proxy
      description: Update an existing LLM proxy
      operationId: updateLLMProxy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLLMProxyRequest'
      responses:
        '200':
          description: Proxy updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LLMProxyResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Proxy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - LLM Proxies
      summary: Delete LLM proxy
      description: Delete an LLM proxy
      operationId: deleteLLMProxy
      responses:
        '204':
          description: Proxy deleted successfully
        '400':
          description: Invalid proxy ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Proxy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================
  # LLM DEPLOYMENT ENDPOINTS
  # ============================================
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Human-readable error message
        description:
          type: string
          description: Error description
        additionalData:
          type: object
          additionalProperties: true
      required:
        - message
...
    UpdateLLMProviderRequest:
      type: object
      minProperties: 1
      properties:
        description:
          type: string
          description: Updated description
        templateUuid:
          type: string
          format: uuid
          description: Updated template UUID
        openapi:
          type: string
          description: Updated OpenAPI specification
        modelList:
          type: array
          description: Updated model list
          items:
            $ref: '#/components/schemas/LLMModelProvider'
        configuration:
          $ref: '#/components/schemas/LLMProviderConfig'
    LLMProviderResponse:
      type: object
      required:
        - uuid
        - templateUuid
        - status
        - configuration
      properties:
        uuid:
          type: string
          format: uuid
          description: UUID of the provider
        description:
          type: string
          description: Provider description
        createdBy:
          type: string
          description: User who created the provider
        templateUuid:
          type: string
          format: uuid
          description: Template UUID
        openapi:
          type: string
          description: OpenAPI specification
        modelProviders:
          type: array
          description: List of model providers
          items:
            $ref: '#/components/schemas/LLMModelProvider'
        status:
          type: string
          description: Provider status
        configuration:
          $ref: '#/components/schemas/LLMProviderConfig'
        artifact:
          $ref: '#/components/schemas/Artifact'
...
    LLMProviderConfig:
      type: object
      properties:
        name:
          type: string
          description: Configuration name
          example: openai-gpt4
        version:
          type: string
          description: Configuration version
          example: v1
        context:
          type: string
          description: API context path
          example: /openai
        vhost:
          type: string
          description: Virtual host
          example: api.example.com
        template:
          type: string
          description: Template name
        upstream:
          $ref: '#/components/schemas/UpstreamConfig'
        accessControl:
          $ref: '#/components/schemas/LLMAccessControl'
        rateLimiting:
          $ref: '#/components/schemas/LLMRateLimitingConfig'
        policies:
          type: array
          items:
            $ref: '#/components/schemas/LLMPolicy'
        security:
          $ref: '#/components/schemas/SecurityConfig'

    # ============================================
    # LLM PROXY SCHEMAS
    # ============================================
...
    UpdateLLMProxyRequest:
      type: object
      minProperties: 1
      properties:
        description:
          type: string
          description: Updated description
        providerUuid:
          type: string
          format: uuid
          description: Updated provider UUID
        openapi:
          type: string
          description: Updated OpenAPI specification
        configuration:
          $ref: '#/components/schemas/LLMProxyConfig'
    LLMProxyResponse:
      type: object
      required:
        - uuid
        - projectId
        - providerUuid
        - status
        - configuration
      properties:
        uuid:
          type: string
          format: uuid
          description: UUID of the proxy
        projectId:
          type: string
          format: uuid
          description: Project UUID
        description:
          type: string
          description: Proxy description
        createdBy:
          type: string
          description: User who created the proxy
        providerUuid:
          type: string
          format: uuid
          description: Provider UUID
        openapi:
          type: string
          description: OpenAPI specification
        status:
          type: string
          description: Proxy status
        configuration:
          $ref: '#/components/schemas/LLMProxyConfig'
        artifact:
          $ref: '#/components/schemas/Artifact'
...
    LLMProxyConfig:
      type: object
      properties:
        name:
          type: string
          description: Proxy name
        version:
          type: string
          description: Proxy version
        context:
          type: string
          description: API context path
        vhost:
          type: string
          description: Virtual host
        provider:
          type: string
          description: Provider reference
        policies:
          type: array
          items:
            $ref: '#/components/schemas/LLMPolicy'
        security:
          $ref: '#/components/schemas/SecurityConfig'

    # ============================================
    # NESTED LLM CONFIGURATION SCHEMAS
    # ============================================
    LLMModelProvider:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          description: Provider ID
          example: openai
        name:
          type: string
          description: Provider name
          example: OpenAI
        models:
          type: array
          items:
            $ref: '#/components/schemas/LLMModel'
    LLMModel:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          description: Model ID
          example: gpt-4
        name:
          type: string
          description: Model name
          example: GPT-4
        description:
          type: string
          description: Model description
          example: Most capable GPT-4 model
...
    UpstreamConfig:
      type: object
      properties:
        main:
          $ref: '#/components/schemas/UpstreamEndpoint'
        sandbox:
          $ref: '#/components/schemas/UpstreamEndpoint'
    UpstreamEndpoint:
      type: object
      properties:
        url:
          type: string
          description: Upstream endpoint URL
          example: https://api.openai.com
        ref:
          type: string
          description: Reference to endpoint configuration
        auth:
          $ref: '#/components/schemas/UpstreamAuth'
    UpstreamAuth:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          description: Authentication type
          enum:
            - apiKey
            - bearer
            - basic
            - none
          example: bearer
        header:
          type: string
          description: Authentication header name
          example: Authorization
        value:
          type: string
          description: Authentication value/token
          example: Bearer sk-...
    LLMAccessControl:
      type: object
      required:
        - mode
      properties:
        mode:
          type: string
          description: Access control mode
          enum:
            - allow
            - deny
          example: allow
        exceptions:
          type: array
          items:
            $ref: '#/components/schemas/RouteException'
    RouteException:
      type: object
      required:
        - path
        - methods
      properties:
        path:
          type: string
          description: Route path pattern
          example: /health
        methods:
          type: array
          items:
            type: string
          description: HTTP methods
          example:
            - GET
            - POST
    LLMPolicy:
      type: object
      required:
        - name
        - version
        - paths
      properties:
        name:
          type: string
          description: Policy name
          example: rate-limiter
        version:
          type: string
          description: Policy version
          example: v1
        paths:
          type: array
          items:
            $ref: '#/components/schemas/LLMPolicyPath'
    LLMPolicyPath:
      type: object
      required:
        - path
        - methods
      properties:
        path:
          type: string
          description: Path pattern
          example: /v1/chat/completions
        methods:
          type: array
          items:
            type: string
          example:
            - POST
        params:
          type: object
          additionalProperties: true
          description: Policy parameters
    LLMRateLimitingConfig:
      type: object
      properties:
        providerLevel:
          $ref: '#/components/schemas/RateLimitingScopeConfig'
        consumerLevel:
          $ref: '#/components/schemas/RateLimitingScopeConfig'
    RateLimitingScopeConfig:
      type: object
      properties:
        global:
          $ref: '#/components/schemas/RateLimitingLimitConfig'
        resourceWise:
          $ref: '#/components/schemas/ResourceWiseRateLimitingConfig'
    ResourceWiseRateLimitingConfig:
      type: object
      required:
        - default
        - resources
      properties:
        default:
          $ref: '#/components/schemas/RateLimitingLimitConfig'
        resources:
          type: array
          items:
            $ref: '#/components/schemas/RateLimitingResourceLimit'
    RateLimitingResourceLimit:
      type: object
      required:
        - resource
        - limit
      properties:
        resource:
          type: string
          description: Resource identifier
          example: /v1/chat/completions
        limit:
          $ref: '#/components/schemas/RateLimitingLimitConfig'
    RateLimitingLimitConfig:
      type: object
      properties:
        request:
          $ref: '#/components/schemas/RequestRateLimit'
        token:
          $ref: '#/components/schemas/TokenRateLimit'
        cost:
          $ref: '#/components/schemas/CostRateLimit'
    RequestRateLimit:
      type: object
      required:
        - enabled
        - count
        - reset
      properties:
        enabled:
          type: boolean
          description: Whether request rate limiting is enabled
        count:
          type: integer
          description: Maximum number of requests
          example: 100
        reset:
          $ref: '#/components/schemas/RateLimitResetWindow'
    TokenRateLimit:
      type: object
      required:
        - enabled
        - count
        - reset
      properties:
        enabled:
          type: boolean
          description: Whether token rate limiting is enabled
        count:
          type: integer
          description: Maximum number of tokens
          example: 100000
        reset:
          $ref: '#/components/schemas/RateLimitResetWindow'
    CostRateLimit:
      type: object
      required:
        - enabled
        - amount
        - reset
      properties:
        enabled:
          type: boolean
          description: Whether cost rate limiting is enabled
        amount:
          type: number
          format: double
          description: Maximum cost amount
          example: 100.0
        reset:
          $ref: '#/components/schemas/RateLimitResetWindow'
    RateLimitResetWindow:
      type: object
      required:
        - duration
        - unit
      properties:
        duration:
          type: integer
          description: Reset window duration
          example: 1
        unit:
          type: string
          description: Time unit
          enum:
            - second
            - minute
            - hour
            - day
          example: hour
    SecurityConfig:
      type: object
      properties:
        enabled:
          type: boolean
          description: Whether security is enabled
          example: true
        apiKey:
          $ref: '#/components/schemas/APIKeySecurity'
    APIKeySecurity:
      type: object
      properties:
        enabled:
          type: boolean
          description: Whether API key security is enabled
        key:
          type: string
          description: API key parameter name
          example: api-key
        in:
          type: string
          description: Where the API key is located
          enum:
            - header
            - query
            - cookie
          example: header
    Artifact:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
          description: Artifact UUID
        name:
          type: string
          description: Artifact name
        displayName:
          type: string
          description: Artifact display name
        description:
          type: string
          description: Artifact description
        status:
          type: string
          description: Artifact status